requirements.txt
Flask
Werkzeug
SpeechRecognition
nltk
numpy
opencv-python
pandas
matplotlib
requests
Pillow
moviepy==1.0.3
gTTS
pydub
python-dotenv
openai
reportlab
pyttsx3
pydantic>=2.0.0
PyYAML>=6.0


